+++ CHANGELOG.md
@@ -17,6 +17,7 @@ HumHub Changelog
 - Fix #6296: Update style variable `background-color-secondary`
 - Fix #6228: Administrators allowed to manage users cannot invite by email or by link if disabled in the settings
 - Fix #6293: Delete content from DB completely on purge content by cron job
+- Fix #39: Integrity Checker should use hardDelete
 
 1.14.0 (April 20, 2023)
 -----------------------